% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiagram.R
\name{add_diagram_spec}
\alias{add_diagram_spec}
\title{Add a compiled diagram spec onto a diagram canvas}
\usage{
add_diagram_spec(plot, spec)
}
\arguments{
\item{plot}{A \code{ggdiagram()} plot.}

\item{spec}{A diagram scene specification.}
}
\value{
A ggplot object with the rendered diagram scene attached.
}
\description{
Add a compiled diagram spec onto a diagram canvas
}
