% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_methods.R
\name{artifacts}
\alias{artifacts}
\title{List recorded session artifacts}
\usage{
artifacts(x, ...)
}
\arguments{
\item{x}{A \code{ggai_session}.}

\item{...}{Passed to methods.}
}
\value{
A list of recorded session artifacts.
}
\description{
List recorded session artifacts
}
