% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_inspection.R
\name{edit_spec}
\alias{edit_spec}
\title{Edit a compiled ggai spec}
\usage{
edit_spec(x, edit, ...)
}
\arguments{
\item{x}{A compiled spec, lazy request, or plot with stored ggai history.}

\item{edit}{A function taking a spec list and returning a modified spec, or a named patch list.}

\item{...}{Passed to methods.}
}
\description{
Edit a compiled ggai spec
}
