% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tools.R
\name{ggai_check_package}
\alias{ggai_check_package}
\title{Check whether an R package is available}
\usage{
ggai_check_package(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
A package availability record.
}
\description{
Check whether an R package is available
}
