% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ggai_default_aisdk_path}
\alias{ggai_default_aisdk_path}
\title{Default local \code{aisdk} development path}
\usage{
ggai_default_aisdk_path()
}
\value{
A scalar character string.
}
\description{
Default local \code{aisdk} development path
}
