% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tools.R
\name{ggai_inspect_examples}
\alias{ggai_inspect_examples}
\title{Inspect local examples for an R help topic}
\usage{
ggai_inspect_examples(package, topic = package)
}
\arguments{
\item{package}{Package name.}

\item{topic}{Example topic.}
}
\value{
A local examples record.
}
\description{
Inspect local examples for an R help topic
}
