% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tools.R
\name{ggai_inspect_help}
\alias{ggai_inspect_help}
\title{Inspect local R help for a package topic}
\usage{
ggai_inspect_help(package, topic = package, max_lines = 80)
}
\arguments{
\item{package}{Package name.}

\item{topic}{Help topic. Defaults to the package name.}

\item{max_lines}{Maximum help lines to return.}
}
\value{
A local help record.
}
\description{
Inspect local R help for a package topic
}
