% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_tools.R
\name{ggai_vignette_index}
\alias{ggai_vignette_index}
\title{List local vignette metadata for an R package}
\usage{
ggai_vignette_index(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
A vignette index record.
}
\description{
List local vignette metadata for an R package
}
