% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrypoints.R
\name{ggdiagram}
\alias{ggdiagram}
\title{Create a diagram canvas}
\usage{
ggdiagram(
  width = 10,
  height = 6,
  background = "white",
  theme = ggai_diagram_theme()
)
}
\arguments{
\item{width}{Canvas width.}

\item{height}{Canvas height.}

\item{background}{Background color.}

\item{theme}{Theme preset name.}
}
\description{
Create a diagram canvas
}
