% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ai_add.R
\name{ggplot_add.ggai_layer_request}
\alias{ggplot_add.ggai_layer_request}
\title{Add a lazy ggai request to an existing ggplot}
\usage{
\method{ggplot_add}{ggai_layer_request}(object, plot, ...)
}
\arguments{
\item{object}{A \code{ggai_layer_request}.}

\item{plot}{A ggplot object.}

\item{...}{Unused compatibility arguments.}
}
\description{
Add a lazy ggai request to an existing ggplot
}
