% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE207092_BrainDev_RNAseq_TPM_anno}
\alias{GSE207092_BrainDev_RNAseq_TPM_anno}
\title{Mouse Brain Development RNA-seq TPM Matrix}
\format{
A compressed CSV file in \verb{data/} with gene annotations and TPM values.
}
\source{
GEO accession GSE207092.
}
\description{
A gzipped CSV copy of the GSE207092 mouse brain-development RNA-seq TPM
annotation table used by the brain-development demos and case study.
}
