% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact.R
\name{is_ggai_artifact}
\alias{is_ggai_artifact}
\title{Test if an object is a \code{ggai_artifact}}
\usage{
is_ggai_artifact(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Test if an object is a \code{ggai_artifact}
}
