% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_methods.R
\name{plot.ggai_session}
\alias{plot.ggai_session}
\title{Plot method for a ggai session}
\usage{
\method{plot}{ggai_session}(x, ...)
}
\arguments{
\item{x}{A \code{ggai_session}.}

\item{...}{Passed through to \code{print.ggplot()}.}
}
\description{
Plot method for a ggai session
}
