% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ggai_polished_figure_result}
\alias{print.ggai_polished_figure_result}
\title{Print a polished figure result summary}
\usage{
\method{print}{ggai_polished_figure_result}(x, ...)
}
\arguments{
\item{x}{A \code{ggai_polished_figure_result}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a polished figure result summary
}
