% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_module.R
\name{render_layout}
\alias{render_layout}
\title{Render a composed plot layout}
\usage{
render_layout(x, ...)
}
\arguments{
\item{x}{A \code{ggai_layout_spec}.}

\item{...}{Unused.}
}
\value{
A grob.
}
\description{
Render a composed plot layout
}
