% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_contracts.R
\name{serialize_ggai_contract}
\alias{serialize_ggai_contract}
\title{Serialize a ggai runtime contract}
\usage{
serialize_ggai_contract(x, pretty = FALSE)
}
\arguments{
\item{x}{A ggai contract object.}

\item{pretty}{Whether to pretty-print JSON.}
}
\value{
A JSON string.
}
\description{
Serialize a ggai runtime contract
}
