% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_edit.R
\name{start_ggai_session}
\alias{start_ggai_session}
\title{Start a stateful ggai editing session}
\usage{
start_ggai_session(plot)
}
\arguments{
\item{plot}{A base ggplot object.}
}
\value{
A \code{ggai_session} object.
}
\description{
Start a stateful ggai editing session
}
