% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_edit.R
\name{undo}
\alias{undo}
\title{Undo the last session edit}
\usage{
undo(session)
}
\arguments{
\item{session}{A \code{ggai_session}.}
}
\value{
The updated \code{ggai_session}.
}
\description{
Undo the last session edit
}
