% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_diagram.R
\name{z_ggai_diagram_spec}
\alias{z_ggai_diagram_spec}
\title{Schema for a compiled diagram scene spec}
\usage{
z_ggai_diagram_spec()
}
\value{
A \code{z_schema} object.
}
\description{
Schema for a compiled diagram scene spec
}
