% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_figure.R
\name{z_ggai_figure_prompt_spec}
\alias{z_ggai_figure_prompt_spec}
\title{Schema for a compiled direct-figure prompt bundle}
\usage{
z_ggai_figure_prompt_spec()
}
\value{
A \code{z_schema} object.
}
\description{
Schema for a compiled direct-figure prompt bundle
}
