% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_glyph.R
\name{z_ggai_glyph_spec}
\alias{z_ggai_glyph_spec}
\title{Schema for a generated glyph spec}
\usage{
z_ggai_glyph_spec()
}
\value{
A \code{z_schema} object.
}
\description{
Schema for a generated glyph spec
}
